changed int to double for dailyDistance

+++ record.hpp
@@ -5,7 +5,7 @@ struct Record {
     std::string ID;
     int Age;
     std::string modeOfTransport;
-    int dailyDistance;
+    double dailyDistance;
     double carbonEmission;
     int monthlyFrequency;
     double monthlyCarbonEmission;
